feat: add tool page for changing old keys, add role to navigation, move beta/connect to connect
